Connect things to docroot as early as possible. (#216)
